backend/requirements.txt
python-dotenv
google-cloud-firestore
google-generativeai
functions-framework